Back-end/requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
groq
python-dotenv
pypdf
numpy
psycopg2